refactored formatting plugin
